Make Add-PacAccessControlEntry2
This function requires a condition, and only works with discretionary (access)
ACEs. You must pass it a security descriptor object (so you can't pass it a
file/folder object or a path).

There is an example of using Get-Acl and Get-PacSecurityDescriptor included in
the comment based help.

The function can later be extended to work with more -InputObject types, but
this should be a good start.