Remove New-PacAccessControlEntry2
Since this is no longer needed, simply remove it. Since we're working around
the limitations of the PAC module, it's better to keep the number of extra
commands to a minimum. For that reason, I don't want to use a proxy function
for New-PacAccessControlEntry since it would rarely be needed (I was
originally using it inside Add-PacAccessControlEntry2, and that function has
been modified to not need it)